+++ test.py
@@ -20,7 +20,7 @@ max_steps = 0
 obs, info = env.reset(seed=1)
 hidden = torch.zeros(15, device=device)
 for _ in range(5000):
-    action, hidden = model(torch.tensor(obs.reshape(SIZE ** 2 + 1), device=device), hidden)
+    action, hidden = model(torch.tensor(obs.reshape(SIZE ** 2 + 1), device=device, dtype=torch.float32), hidden)
     obs, reward, terminated, truncated, info = env.step(torch.argmax(action))
 
     if terminated:
